UnitTestRunner:  Fixed some build warnings on Linux

+++ modules/juce_gui_basics/native/juce_linux_X11_Windowing.cpp
@@ -2216,7 +2216,7 @@ public:
 
         auto mapIndex = (uint32) (buttonPressEvent.button - Button1);
 
-        if (mapIndex < numElementsInArray (pointerMap))
+        if (mapIndex < (uint32) numElementsInArray (pointerMap))
         {
             switch (pointerMap[mapIndex])
             {
@@ -2241,7 +2241,7 @@ public:
 
         auto mapIndex = (uint32) (buttonRelEvent.button - Button1);
 
-        if (mapIndex < numElementsInArray (pointerMap))
+        if (mapIndex < (uint32) numElementsInArray (pointerMap))
         {
             switch (pointerMap[mapIndex])
             {
